SDK-3654. Fix type warning as error in integration test

Closes SDK-3654

See merge request sdk/sdk!5302

+++ tests/integration/SdkTest_test.cpp
@@ -17158,8 +17158,8 @@ TEST_F(SdkTest, GiveRemoveChatAccess)
 
     // Create chat between new contacts
 
-    int numChatsHost = mApi[host].chats.size();
-    int numChatsGuest = mApi[guest].chats.size();
+    long unsigned int numChatsHost = mApi[host].chats.size();
+    long unsigned int numChatsGuest = mApi[guest].chats.size();
     mApi[guest].chatUpdated = false;
     std::unique_ptr<MegaTextChatPeerList> peers(MegaTextChatPeerList::createInstance());
     peers->addPeer(megaApi[guest]->getMyUser()->getHandle(), PRIV_STANDARD);
